feat: no era con Long, era con Double

+++ desarrollo/src/main/java/com/example/desarrollo/dto/HabitacionResponseDTO.java
@@ -15,8 +15,8 @@ public class HabitacionResponseDTO {
 
     @NotNull
     private String direccion;
-    private Long longitud;
-    private Long latitud;
+    private Double longitud;
+    private Double latitud;
 
     @NotNull
     private Integer area;
